requirements.txt
PyQt5
python-nmap
